+++ src/qsolve/citardarqFormula.h
@@ -1,1 +1,1 @@
-float * citardarqFormula(double aVal,double bVal,double cVal);
+double * citardarqFormula(double aVal,double bVal,double cVal);
